[Refactor] Move stitched components into files
