+++ src/desktop/components/react/stitch_components/ReactionArtwork.tsx
@@ -1,0 +1,12 @@
+import React from "react"
+import { ContextProvider } from "reaction/Artsy"
+import { data as sd } from "sharify"
+import { Artwork } from "reaction/Components/Artwork"
+
+export const ReactionArtwork = props => {
+  return (
+    <ContextProvider user={sd.CURRENT_USER}>
+      <Artwork {...props} useRelay={false} />
+    </ContextProvider>
+  )
+}
