+++ src/desktop/components/react/stitch_components/utils/mapToRelayConnection.ts
@@ -1,0 +1,12 @@
+/**
+ * Ensure that old artwork collections conform to new Relay connection api
+ */
+export const mapToRelayConnection = artworks => {
+  return {
+    edges: artworks.map(artwork => {
+      return {
+        node: artwork,
+      }
+    }),
+  }
+}
