+++ src/desktop/lib/global_client_setup.coffee
@@ -19,7 +19,7 @@ setupSplitTests = require '../components/split_test/setup.coffee'
 listenForInvert = require '../components/eggs/invert/index.coffee'
 listenForBounce = require '../components/eggs/bounce/index.coffee'
 confirmation = require '../components/confirmation/index.coffee'
-globalReactModules = require('./global_react_modules')
+globalReactModules = require('./global_react_modules.tsx')
 { componentRenderer } = require('@artsy/stitch/iso')
 { initModalManager } = require('../../desktop/apps/auth2/client/index')
