+++ src/desktop/lib/global_react_modules.tsx
@@ -1,0 +1,1 @@
+export * from "desktop/components/react/stitch_components"
